swarm/netsim: move http related things into separate files
